# swagger.yaml
openapi: 3.0.3
info:
  title: Friendface-API v1
  version: 0.1.0
  description: |-
    This API is used for [Friendface](https://friendface.vercel.app/) social media project with use case of doing CRUD to database and store image file. 

    The API handle some social media schema including post, comment, like, user, userChat, messages, notification, messages notification, and chat room.
    
    In general the data have two different category. One for social media data and the other one is of the messager feature.
    
    _Some related link:_
    - [Friendface-client repository](https://github.com/kurniadikevin/friendface-client)
    - [Friendface App](https://friendface.vercel.app/)
    - [Developer](https://github.com/kurniadikevin)
servers:
  - url: http://localhost:5000
...
paths:
  /posts:
    get:
      tags:
        - posts
      summary: Get all post
      description: Get all post data on one endpoint
      operationId: getAllPosts
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema: 
                type: array       
        '400':
          description: Invalid 
...
  /users:
    get:
      tags:
        - users
      summary: Get all user data
      description: Get all post data on one endpoint
      operationId: getAllUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Users'       
        '400':
          description: Invalid 
...
  /users/login:
    post:
      tags:
        - users
      summary: Login user
      description: Login user with email and password
      operationId: loginUser
      requestBody:
        description: Log in new user
        content:
          application/json:
             schema:
                type: object
                properties:
                  email:
                    type: string
                  password:
                    type: string
        required: true
      responses:
        '200':
          description: Successful operation user logged-in
          content:
            application/json:
        '400':
          description: Invalid
...
  /images:
    post:
      tags:
        - images
      summary: Upload new image for profile picture
      description: Upload new image for profile picture
      operationId: newImageProfilePicture
      responses:
        '200':
          description: Successful operation message created     
        '400':
          description: Invalid
      parameters:
        - name: chatRoomId
          in: path
          description: target chatRoomId
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
             schema:
                type: object
                properties:
                  _id:
                    type: string
                    description: id of user that update profile picture
                  byUser:
                    type: string
                    description: current user, user id target
                    example: 6364fff02066d5b0ca6ddb7d
                  name:
                    type: string
                  desc:
                    type: string
                  url:
                    type: string
                    description: url of image 
                  fileName:
                    type: string
                    format: binary
  /postImages:
    post:
      tags:
        - images 
        - posts
      summary: Upload new image for post
      description: Upload new image for post with text content for post
      operationId: newImagePost
      responses:
        '200':
          description: Successful operation message created     
        '400':
          description: Invalid
      requestBody:
        content:
           application/json:
             schema:
                type: object
                properties:
                 authorId:
                    type: string
                    description: current user, user id target
                    example: 6364fff02066d5b0ca6ddb7d
                 text:
                    type: string
           multipart/form-data:
             schema:
                type: object
                properties:
                 fileName:
                    type: string
                    format: binary
components:
  schemas:
...
    Users:
      required:
        - email
        - password
      type: object
      properties:
        _id:
          type: string
          example: 6374e91e5ffbe2e45ddf4045
        username:
          type: string
          example: TestAccount
        email:
          type: string
          example: guest@gmail.com
        password:
          type: string
          example: 123456
        friends:
          type: array
          example: [{_id: "6374e91e5ffbe2e45ddf4045", email: "guest@gmail.com", username: "TestAccount",…}]
        profilePicture:
          type: string
          description: url for profile picture
        friendRequest:
          type: array
          example: [{ seenAt: '1681281189976', sender: {_id: "6414298aa9d5395dc8c79c6a"},status: "pending"}]
        postNotification:
          type: array
          example: [{action: "Liked", byUser : {_id: "6367b17ab6704e28258dddf4"}, postId : "6419ab32b8a9bb03162c3968", seenAt: '1681306103830'}]
...